[controller] Extract VersionLifecyclePolicy from VeniceHelixAdmin / VeniceParentHelixAdmin
First lifecycle helper in the new versionlifecycle sub-package: a pure-statics
class (final, no I/O, no admin coupling) holding the new-push capacity guards,
version-selection helpers, push-status aggregation, and TTL repush flag rule.
Mirrors HybridStoreConfigPolicy's shape and is the first of five planned
helpers
(TopicLifecycleManager, VersionRetirementManager, VersionProvisioner,
CurrentVersionManager) for the versionlifecycle extraction.

Lifted (with call-site rewires):
  - checkBackupVersionCleanupCapacityForNewPush
  - checkRollbackOriginVersionCapacityForNewPush
  - getStartedVersion
  - getBackupVersionNumber  (also called from VeniceParentHelixAdmin +
DataRecoveryManager)
  - hasFatalDataValidationError
  - updateStoreTTLRepushFlag
  - getFinalReturnStatus  (from VeniceParentHelixAdmin)
  - getOverallPushStatus
  - STATUS_PRIORITIES constant

Method bodies are byte-identical to the originals; call sites in
VeniceHelixAdmin
inline the (config, time) args that the deleted instance wrappers used to
resolve.

Tests: VersionLifecyclePolicyTest consolidates the two standalone test files
(TestBackupVersionCleanupCapacityCheck,
TestRollbackOriginVersionCapacityCheck)
plus the in-admin tests for getFinalReturnStatus / getOverallPushStatus /
updateStoreTTLRepushFlag, and adds fresh coverage for getStartedVersion,
getBackupVersionNumber, hasFatalDataValidationError. Full
:services:venice-controller:test passes.

Net: VeniceHelixAdmin -207, VeniceParentHelixAdmin -52; new policy + test
land at
311 + 591 lines.
