+++ services/venice-controller/src/test/java/com/linkedin/venice/controller/TestVeniceParentHelixAdmin.java
@@ -3431,86 +3431,6 @@ public class TestVeniceParentHelixAdmin extends AbstractTestVeniceParentHelixAdm
   }
 
   @Test
-  public void testGetFinalReturnStatus() {
-    Map<String, ExecutionStatus> statuses = new HashMap<>();
-    Set<String> childRegions = new HashSet<>();
-    childRegions.add("region1");
-    childRegions.add("region2");
-    childRegions.add("region3");
-    ExecutionStatus finalStatus;
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.COMPLETED);
-    statuses.put("region3", ExecutionStatus.COMPLETED);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.COMPLETED);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.PROGRESS);
-    statuses.put("region3", ExecutionStatus.COMPLETED);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.PROGRESS);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.ERROR);
-    statuses.put("region3", ExecutionStatus.COMPLETED);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.ERROR);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.ERROR);
-    statuses.put("region3", ExecutionStatus.UNKNOWN);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 1, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.UNKNOWN);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.UNKNOWN);
-    statuses.put("region2", ExecutionStatus.ERROR);
-    statuses.put("region3", ExecutionStatus.UNKNOWN);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 2, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.PROGRESS);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.COMPLETED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.COMPLETED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.DVC_INGESTION_ERROR_MEMORY_LIMIT_REACHED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.DVC_INGESTION_ERROR_DISK_FULL);
-
-    statuses.clear();
-    statuses.put("region1", ExecutionStatus.COMPLETED);
-    statuses.put("region2", ExecutionStatus.DVC_INGESTION_ERROR_MEMORY_LIMIT_REACHED);
-    statuses.put("region3", ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-    finalStatus = VeniceParentHelixAdmin.getFinalReturnStatus(statuses, childRegions, 0, new StringBuilder());
-    assertEquals(finalStatus, ExecutionStatus.DVC_INGESTION_ERROR_OTHER);
-  }
-
-  @Test
   public void testUpdateAdminOperationProtocolVersion() {
     String clusterName = "test-cluster";
     Long adminProtocolVersion = 10L;
